+++ example/android/app/src/main/kotlin/com/pavlok/flutter_system_ringtones_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.pavlok.flutter_system_ringtones_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}